Fix memory leak
I was watching [the Elixir Sips episode 174][E174] and noticed the use
of `Enum.to_list/1` to trigger enumeration. That seems problematic as
the list will have to contain an unconstrained number of items.

I replaced this with `Stream.run/1` which "is useful when a stream needs
to be run, for side effects, and there is no interest in its return
result".

As `ImageTweetStreamer.stream/1` is responsible for processing tweets,
maybe the better design is to not make it lazy.

[E174]: http://elixirsips.com/episodes/174_elixir_friends_saving_tweets_with_streams_and_filters.html